+++ source/ftrack_connect/ui/application.py
@@ -287,11 +287,18 @@ class Application(QtWidgets.QMainWindow):
         if sys.platform == "darwin":
             from darkdetect import _mac_detect as stylemodule
 
-        elif sys.platform == "win32" and int(platform.release()) >= 10:
+        elif (
+                sys.platform == "win32"
+                and platform.release().isdigit()  # Windows 8.1 would result in '8.1' str
+                and int(platform.release()) >= 10
+        ):
             from darkdetect import _windows_detect as stylemodule
 
         elif sys.platform == "linux":
             from darkdetect import _linux_detect as stylemodule
+
+        else:
+            from darkdetect import _dummy as stylemodule
 
         current_theme = stylemodule.theme()
         if not current_theme:
@@ -764,7 +771,7 @@ class Application(QtWidgets.QMainWindow):
         for apiPluginPath in os.environ.get(
             'FTRACK_EVENT_PLUGIN_PATH', ''
         ).split(os.pathsep):
-            if apiPluginPath not in plugin_paths:
+            if apiPluginPath and apiPluginPath not in plugin_paths:
                 plugin_paths.append(os.path.expandvars(apiPluginPath))
 
         for connectPluginPath in os.environ.get(
@@ -774,13 +781,15 @@ class Application(QtWidgets.QMainWindow):
                 self._gatherPluginHooks(os.path.expandvars(connectPluginPath))
             )
 
+        plugin_paths = list(set(plugin_paths))
+
         self.logger.info(
             u'Connect plugin hooks directories: {0}'.format(
                 ', '.join(plugin_paths)
             )
         )
 
-        return list(plugin_paths)
+        return plugin_paths
 
     def _gatherPluginHooks(self, path):
         '''Return plugin hooks from *path*.'''