+++ source/ftrack_connect/ui/widget/login.py
@@ -1,5 +1,7 @@
 # :coding: utf-8
 # :copyright: Copyright (c) 2014 ftrack
+import os
+import logging
 
 from ftrack_connect.qt import QtWidgets, QtCore, QtGui
 
@@ -26,8 +28,11 @@ class Login(QtWidgets.QWidget):
     def __init__(self, *args, **kwargs):
         '''Instantiate the login widget.'''
         super(Login, self).__init__()
-
+        self.logger = logging.getLogger(
+            __name__ + '.' + self.__class__.__name__
+        )
         theme = kwargs.get("theme", "light")
+        self.server_urls = kwargs.get("server_urls", None)
 
         layout = QtWidgets.QVBoxLayout()
         layout.addSpacing(50)
@@ -56,9 +61,15 @@ class Login(QtWidgets.QWidget):
         label.setObjectName('login-label')
         layout.addWidget(label, alignment=QtCore.Qt.AlignCenter)
 
+        if os.getenv('FTRACK_SERVER', ''):
+            self.logger.warning('FTRACK_SERVER '
+                                'environment variable is set to {}'.format(
+                os.getenv('FTRACK_SERVER', '')))
+
         self.server = QtWidgets.QLineEdit()
         self.server.setPlaceholderText('Site name or custom domain URL')
         layout.addWidget(self.server)
+        self.server.setText(os.getenv('FTRACK_SERVER', ''))
 
         self.username = QtWidgets.QLineEdit()
         self.username.setPlaceholderText('User name')
@@ -105,7 +116,7 @@ class Login(QtWidgets.QWidget):
         self.toggle_api_label.setObjectName('lead-label')
         self.toggle_api_label.setText(
             'Trouble signing in? '
-            '<a href="#" style="color: #FFDD86;">Sign in with username and API key</a>'
+            '<a href="#" style="color: #FFDD86;">Sign in with username and API key.</a>'
         )
         self.toggle_api_label.clicked.connect(self._toggle_credentials)
         layout.addWidget(
@@ -124,6 +135,17 @@ class Login(QtWidgets.QWidget):
             self.untoggle_api_label, alignment=QtCore.Qt.AlignCenter
         )
         layout.addSpacing(20)
+
+        if self.server_urls:
+            self.logger.debug('Setting up Completer for server urls')
+            self.server_completer = QtWidgets.QCompleter(
+                list(self.server_urls.keys()), self)
+            self.server_completer.setCompletionMode(
+                QtWidgets.QCompleter.PopupCompletion)
+            self.server.setCompleter(self.server_completer)
+            self.server_completer.activated.connect(self.server_url_chosen)
+            if os.getenv('FTRACK_SERVER', ''):
+                self.server_url_chosen()
 
 
 
@@ -151,3 +173,35 @@ class Login(QtWidgets.QWidget):
         self.username.hide()
         self.toggle_api_label.show()
         self.untoggle_api_label.hide()
+
+    def server_url_chosen(self):
+        self._toggle_credentials()
+        credentials = self.server_urls.get(self.server.text(), None)
+        if credentials:
+            self.user_names = {}
+            for i in credentials["accounts"]:
+                self.user_names[i['api_user']] = i['api_key']
+                if i['api_user'] == credentials['last_used_api_user']:
+                    self.username.setText(i['api_user'])
+                    self.apiKey.setText(i['api_key'])
+            if self.user_names:
+                self.logger.debug('Setting up Completer for user names')
+                # this would prevent the user to be able to ever choose
+                # a different login then one that was specified
+                # if len(list(self.user_names.keys())) == 1:
+                #     self.handleLogin()
+                #     return
+                try:
+                    self.user_completer.activated.disconnect(self.user_chosen)
+                except:
+                    pass
+                self.user_completer = QtWidgets.QCompleter(
+                    list(self.user_names.keys()), self)
+                self.user_completer.setCompletionMode(
+                    QtWidgets.QCompleter.PopupCompletion)  # pylint: disable=line-too-long
+                self.username.setCompleter(self.user_completer)
+                self.user_completer.activated.connect(self.user_chosen)
+
+    def user_chosen(self):
+        self.logger.debug('Getting api key from Completer')
+        self.apiKey.setText(self.user_names[self.username.text()])